+++ src/__init__.py
@@ -1,0 +1,10 @@
+"""
+AWS Export Resources - Source Package
+
+This package contains the core AWS resource export functionality.
+"""
+
+from ._version import __version__
+
+__author__ = "Nattait Nandawarang"
+__organization__ = "Gosoft (Thailand) Co., Ltd."